+++ vimlib/frain/__init__.py
@@ -1,8 +1,11 @@
 # -*- coding:utf-8 -*-
 #    author    :   丁雪峰
-#    time      :   2015-04-08 14:10:11
+#    time      :   2015-04-09 11:56:54
 #    email     :   fengidri@yeah.net
 #    version   :   1.0.1
+
+
+from frain import FrainList as LIST
 
 
 if __name__ == "__main__":
