--- vimlib/frainui/data.py
@@ -1,11 +1,0 @@
-# -*- coding:utf-8 -*-
-#    author    :   丁雪峰
-#    time      :   2015-04-08 10:43:15
-#    email     :   fengidri@yeah.net
-#    version   :   1.0.1
-LISTWIN = None
-
-
-if __name__ == "__main__":
-    pass
-
